Shadows

+++ TGC.MonoGame.TP/Sources/Scenes/World.cs
@@ -55,7 +55,7 @@ namespace TGC.MonoGame.TP.Scenes
                 return;
 
             if (random.NextDouble() > 0.8f)
-                new TIE().Instantiate(new Vector3((float)random.Next(-4000, 4000), 0f, 0f));
+                new TIE().Instantiate(new Vector3(random.Next(-4000, 4000), 0f, 0f));
             lastTieSpawn = gameTime.TotalGameTime.TotalMilliseconds;
         }
